chore: CommentType -> CardType
It will be used in the agile plan and voting page
